+++ src/main/java/cart/controller/PageController.java
@@ -1,0 +1,28 @@
+package cart.controller;
+
+import cart.service.ProductService;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class PageController {
+
+    private final ProductService productService;
+
+    public PageController(final ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping("/")
+    public String index(final Model model) {
+        model.addAttribute("products", productService.findAll());
+        return "index";
+    }
+
+    @GetMapping("/admin")
+    public String admin(final Model model) {
+        model.addAttribute("products", productService.findAll());
+        return "admin";
+    }
+}
