+++ src/test/java/cart/controller/ProductControllerTest.java
@@ -1,18 +1,11 @@
 package cart.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.hasItem;
-import static org.hamcrest.Matchers.hasProperty;
-import static org.hamcrest.Matchers.hasSize;
-import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import cart.dao.ProductDao;
@@ -34,7 +27,7 @@ import org.springframework.transaction.annotation.Transactional;
 @DisplayNameGeneration(ReplaceUnderscores.class)
 @AutoConfigureMockMvc
 @Transactional
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@SpringBootTest
 class ProductControllerTest {
 
     @Autowired
@@ -66,66 +59,6 @@ class ProductControllerTest {
                 () -> assertThat(result.getImage()).isEqualTo("tea.jpg"),
                 () -> assertThat(result.getPrice()).isEqualTo(1000L)
         );
-    }
-
-    @Test
-    void 메인_페이지_접근() throws Exception {
-        // given
-        final Product product1 = new Product("허브티", "tea.jpg", 1000L);
-        final Product product2 = new Product("고양이", "cat.jpg", 1000000L);
-        final Long id1 = productDao.saveAndGetId(product1).get();
-        final Long id2 = productDao.saveAndGetId(product2).get();
-
-        // expect
-        mockMvc.perform(get("/"))
-                .andExpect(model().attribute("products", hasSize(2)))
-                .andExpect(model().attribute("products", hasItem(
-                        allOf(
-                                hasProperty("id", is(id1)),
-                                hasProperty("name", is("허브티")),
-                                hasProperty("image", is("tea.jpg")),
-                                hasProperty("price", is(1000L))
-                        )
-                )))
-                .andExpect(model().attribute("products", hasItem(
-                        allOf(
-                                hasProperty("id", is(id2)),
-                                hasProperty("name", is("고양이")),
-                                hasProperty("image", is("cat.jpg")),
-                                hasProperty("price", is(1000000L))
-                        )
-                )))
-                .andDo(print());
-    }
-
-    @Test
-    void 관리자_페이지_접근() throws Exception {
-        // given
-        final Product product1 = new Product("허브티", "tea.jpg", 1000L);
-        final Product product2 = new Product("고양이", "cat.jpg", 1000000L);
-        final Long id1 = productDao.saveAndGetId(product1).get();
-        final Long id2 = productDao.saveAndGetId(product2).get();
-
-        // expect
-        mockMvc.perform(get("/admin"))
-                .andExpect(model().attribute("products", hasSize(2)))
-                .andExpect(model().attribute("products", hasItem(
-                        allOf(
-                                hasProperty("id", is(id1)),
-                                hasProperty("name", is("허브티")),
-                                hasProperty("image", is("tea.jpg")),
-                                hasProperty("price", is(1000L))
-                        )
-                )))
-                .andExpect(model().attribute("products", hasItem(
-                        allOf(
-                                hasProperty("id", is(id2)),
-                                hasProperty("name", is("고양이")),
-                                hasProperty("image", is("cat.jpg")),
-                                hasProperty("price", is(1000000L))
-                        )
-                )))
-                .andDo(print());
     }
 
     @Test
